Product managment: add security check if the user is admin

+++ products/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import get_object_or_404, redirect, render
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+
 from django.urls import reverse
 from django.db.models import Q
 from django.db.models.functions import Lower
@@ -80,8 +82,15 @@ def product_detail(request, product_id):
     return render(request, "products/product_detail.html", context)
 
 
+@login_required
 def add_product(request):
     """Add a product to the store"""
+
+    # only store owners can manage products
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     if request.method == "POST":
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -104,8 +113,15 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """Edit a product in the store"""
+
+    # only store owners can manage products
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
 
     if request.method == "POST":
@@ -132,8 +148,15 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """Delete a product from the store"""
+
+    # only store owners can manage products
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
